API connection in Android Studio

+++ androidApplication/app/src/main/java/utt/student/greenfresh/classes/AreaReadings.java
@@ -3,19 +3,19 @@ package utt.student.greenfresh.classes;
 public class AreaReadings {
     // attributes
     private ProductionLine productionLine;
-    private Float temperature;
+    private double temperature;
     private int humidity;
 
     // getters and setters
     public ProductionLine getProductionLine() { return productionLine; }
     public void setProductionLine(ProductionLine productionLine) { this.productionLine = productionLine; }
-    public Float getTemperature() { return temperature; }
-    public void setTemperature(Float temperature) { this.temperature = temperature; }
+    public double getTemperature() { return temperature; }
+    public void setTemperature(double temperature) { this.temperature = temperature; }
     public int getHumidity() { return humidity; }
     public void setHumidity(int humidity) { this.humidity = humidity; }
 
     // constructors
-    public AreaReadings(ProductionLine productionLine, Float temperature, int humidity) {
+    public AreaReadings(ProductionLine productionLine, double temperature, int humidity) {
         this.productionLine = productionLine;
         this.temperature = temperature;
         this.humidity = humidity;
